Public API Documentation

Documentation for CRTBPNaturalMotion's public interface.

Contents

Contents

Index

Periodic Orbit Computation

Invariant Manifold Computation

Type Flags

Propagation Functions

Utility Functions

Index

CRTBPNaturalMotion.ArcLength

CRTBPNaturalMotion.InvariantManifold

CRTBPNaturalMotion.Time

CRTBPNaturalMotion.TypeAPeriodicOrbit

CRTBPNaturalMotion.TypeAPeriodicOrbit

CRTBPNaturalMotion.TypeAPeriodicOrbit

CRTBPNaturalMotion.correct_typeA_initial_conditions

CRTBPNaturalMotion.generate_periodic_orbit_cheb_approximation

CRTBPNaturalMotion.generate_periodic_orbit_cheb_interpolation

CRTBPNaturalMotion.get_full_orbit

CRTBPNaturalMotion.get_stable_manifold_trajectory

CRTBPNaturalMotion.get_stable_manifold_trajectory

CRTBPNaturalMotion.get_stable_manifold_trajectory

CRTBPNaturalMotion.get_stable_manifold_trajectory

CRTBPNaturalMotion.get_unstable_manifold_trajectory

CRTBPNaturalMotion.get_unstable_manifold_trajectory

CRTBPNaturalMotion.get_unstable_manifold_trajectory

CRTBPNaturalMotion.get_unstable_manifold_trajectory

CRTBPNaturalMotion.jacobi_integral

CRTBPNaturalMotion.jacobian

CRTBPNaturalMotion.jacobian

CRTBPNaturalMotion.propagate_return_all_states

CRTBPNaturalMotion.propagate_return_all_states

CRTBPNaturalMotion.propagate_return_all_states

CRTBPNaturalMotion.propagate_return_all_states

CRTBPNaturalMotion.propagate_return_arc_length

CRTBPNaturalMotion.propagate_return_final_state

CRTBPNaturalMotion.propagate_return_final_state

CRTBPNaturalMotion.propagate_return_final_stm

CRTBPNaturalMotion.propagate_return_time

CRTBPNaturalMotion.value

Periodic Orbit Computation

CRTBPNaturalMotion.TypeAPeriodicOrbit — Type

TypeAPeriodicOrbit

A representation of a CRTBP periodic orbit of Type A.

Fields

u0::SVector{3, Float64}: The initial conditions for the periodic orbit.

P::Float64: The period of the periodic orbit.

S::Float64: The arc length along the periodic orbit.

N_cross::Int: The number of crossings through x-z plane in half period.

mu::Float64: The gravitational parameter of the CRTBP.

CRTBPNaturalMotion.TypeAPeriodicOrbit — Method

TypeAPeriodicOrbit(
    rx, rz, vy, mu;
    N_cross = 1,
    constraint = (:x_start_coordinate, 0.8),
    ode_solver = Vern9(),
    ode_reltol = 1e-12,
    ode_abstol = 1e-12,
    nl_solver  = SimpleTrustRegion(autodiff = nothing, nlsolve_update_rule = Val(true)),
)

CRTBPNaturalMotion.TypeAPeriodicOrbit — Method

TypeAPeriodicOrbit(
    orbit::TypeAPeriodicOrbit;
    constraint = (:x_start_coordinate, 0.8),
    ode_solver = Vern9(),
    ode_reltol = 1e-12,
    ode_abstol = 1e-12,
    nl_solver  = SimpleTrustRegion(autodiff = nothing, nlsolve_update_rule = Val(true)),
)

CRTBPNaturalMotion.generate_periodic_orbit_cheb_approximation — Method

generate_periodic_orbit_cheb_approximation(
    orbit::AbstractPeriodicOrbit, τ1_npoints::Int, τ1_order::Int, PT::ParameterizationType;
    solver = Vern9(),
    reltol = 1e-14,
    abstol = 1e-14,
) -> FastChebInterpolation{FastChebInterp.ChebPoly{1, SVector{6, Float64}, Float64}}

Generate a Chebyshev approximation of the periodic orbit in terms of PT employing τ1_npoints fit to a Chebyshev polynomial of order τ1_order.

Arguments

orbit::AbstractPeriodicOrbit: The periodic orbit.

τ1_npoints::Int: The number of points to fit the Chebyshev polynomial.

τ1_order::Int: The order of the Chebyshev polynomial.

PT::ParameterizationType: The type of parameterization to use (Time or ArcLength).

Keyword Arguments

solver::OrdinaryDiffEq.AbstractODESolver: The ODE solver to use for the integration.

reltol::Real: The relative tolerance for the ODE solver.

abstol::Real: The absolute tolerance for the ODE solver.

Returns

FastChebInterpolation{FastChebInterp.ChebPoly{1, SVector{6, Float64}, Float64}}: The Chebyshev approximation.

CRTBPNaturalMotion.generate_periodic_orbit_cheb_interpolation — Method

generate_periodic_orbit_cheb_interpolation(
    orbit::AbstractPeriodicOrbit, τ1_order::Int, PT::ParameterizationType;
    solver = Vern9(),
    reltol = 1e-14,
    abstol = 1e-14,
) -> FastChebInterpolation{FastChebInterp.ChebPoly{1, SVector{6, Float64}, Float64}}

Generate a Chebyshev interpolation of the periodic orbit in terms of PT employing a Chebyshev polynomial of order τ1_order.

Arguments

orbit::AbstractPeriodicOrbit: The periodic orbit.

τ1_order::Int: The order of the Chebyshev polynomial.

PT::ParameterizationType: The type of parameterization to use (Time or ArcLength).

Keyword Arguments

solver::OrdinaryDiffEq.AbstractODESolver: The ODE solver to use for the integration.

reltol::Real: The relative tolerance for the ODE solver.

abstol::Real: The absolute tolerance for the ODE solver.

Returns

FastChebInterpolation{FastChebInterp.ChebPoly{1, SVector{6, Float64}, Float64}}: The Chebyshev interpolation.

CRTBPNaturalMotion.get_full_orbit — Method

get_full_orbit(orbit::AbstractPeriodicOrbit)

Get the full orbit of the periodic orbit.

Arguments

orbit::AbstractPeriodicOrbit: The periodic orbit.

CRTBPNaturalMotion.correct_typeA_initial_conditions — Method

correct_typeA_initial_conditions(
    rx_guess, rz_guess, vy_guess, mu;
    N_cross     = 1,
    constraint  = (:x_start_coordinate, 0.0),
    ode_solver  = Vern9(),
    ode_reltol  = 1e-14,
    ode_abstol  = 1e-14,
    nl_solver   = SimpleTrustRegion(autodiff = nothing, nlsolve_update_rule = Val(true)),
)

Correct the initial conditions of a type-A periodic orbit using a forward shooting method.

Arguments

rx_guess::Real: Initial guess for the x-coordinate of the initial state.

rz_guess::Real: Initial guess for the z-coordinate of the initial state.

vy_guess::Real: Initial guess for the y-velocity of the initial state.

mu::Real: Gravitational parameter of the CRTBP.

Keyword Arguments

N_cross::Int = 1: Number of crossings of the xz-plane to consider in the half period.

constraint::Tuple{Symbol,Real}: Tuple of the constraint type and value corresponding to the final constraint considered to provide a fully defined system of nonlinear eqautions. The constraint type can be either :x_start_coordinate or :jacobi_integral. The constraint value is the value that the constraint should be equal to for the desired periodic orbit.

ode_solver::OrdinaryDiffEq.AbstractODESolver: The ODE solver to use for the integration.

ode_reltol::Real: The relative tolerance for the ODE solver.

ode_abstol::Real: The absolute tolerance for the ODE solver.

nl_solver::NonlinearSolve.AbstractNLsolveSolver: The nonlinear solver to use for the root-finding problem. Note, there seems to be a strange bug with the NonlinearSolve.jl TrustRegion and NewtonRaphson solvers when using out-of-place Jacobian defined to return an SMatrix. Therefore, it is recommened to use either the default SimpleTrustRegion solver or the SimpleNewtonRaphson solver (which are specifically tailored towards StaticArrays.jl arrays).

Invariant Manifold Computation

CRTBPNaturalMotion.InvariantManifold — Type

InvariantManifold{PO <: AbstractPeriodicOrbit}

A struct for invariant manifolds of periodic orbits.

Fields

orbit::PO: The periodic orbit.

Δr::Float64: The manifold perturbation size (change in position) along stable/unstable eigenvector direction.

CRTBPNaturalMotion.get_stable_manifold_trajectory — Method

get_stable_manifold_trajectory(
    man::InvariantManifold, left_pert::Bool,
    τ1::AbstractFloat, PT1::ParameterizationType,
    start_cond::Union{AbstractFloat,Function},
    τ2::Union{AbstractFloat,ParameterListType}, PT2::ParameterizationType;
    manifold_length::AbstractFloat = 1.0,
    solver = Vern9(),
    reltol = 1e-14,
    abstol = 1e-14,
)

Returns a single stable manifold trajectory corresponding to the parameters τ1 and τ2. Here, the starting point of the returned trajectory is determined by the argument start_cond. If start_cond is a scalar, this specifies the distance from the initial condition on the periodic orbit in terms of PT2. If start_cond is a function, this specifies a termination condition callback function for the solver that will be used to propagate to the initial state of the returned trajectory from the initial condition on the periodic orbit.

Arguments

man::InvariantManifold: The invariant manifold.

left_pert::Bool: If true, the perturbation is to the left of the periodic orbit.

τ1::AbstractFloat: The parameter for the initial condition on the periodic orbit.

PT1::ParameterizationType: The parameterization type for the τ1 parameter.

start_cond::Union{AbstractFloat,Function}: The starting condition for the manifold trajectory. If a scalar, this specifies the distance from the initial condition on the periodic orbit in terms of PT2. If a function, this specifies a termination condition callback function for the solver that will be used to propagate to the initial state of the returned trajectory from the initial condition on the periodic orbit.

τ2::Union{AbstractFloat,ParameterListType}: The parameter for the final condition on the manifold. If a single scalar is provided, the returned trajectory is a DifferentialEquations.jl solution struct. If a list is provided, the returned trajectory is an array of states corresponding to each element in the list.

PT2::ParameterizationType: The parameterization type for the τ2 parameter.

Keyword Arguments

manifold_length::AbstractFloat = 1.0: The length of the manifold, in CRTBP units. Sets the max length of the manifold, where τ2 = 1.0 corresponds to the full manifold_length, starting from the state parameterized by start_cond.

solver = Vern9(): The DifferentialEquations.jl solver to use.

reltol = 1e-14: The relative tolerance for the solver.

abstol = 1e-14: The absolute tolerance for the solver.

CRTBPNaturalMotion.get_stable_manifold_trajectory — Method

get_stable_manifold_trajectory(
    man::InvariantManifold, left_pert::Bool,
    τ1::AbstractFloat, PT1::ParameterizationType,
    τ2::Union{AbstractFloat,ParameterListType}, PT2::ParameterizationType;
    manifold_length::AbstractFloat = 1.0,
    solver = Vern9(),
    reltol = 1e-14,
    abstol = 1e-14,
)

Returns a single stable manifold trajectory corresponding to the parameters τ1 and τ2.

Arguments

man::InvariantManifold: The invariant manifold.

left_pert::Bool: If true, the perturbation is to the left of the periodic orbit.

τ1::AbstractFloat: The parameter for the initial condition on the periodic orbit.

PT1::ParameterizationType: The parameterization type for the τ1 parameter.

τ2::Union{AbstractFloat,ParameterListType}: The parameter for the final condition on the manifold. If a single scalar is provided, the returned trajectory is a DifferentialEquations.jl solution struct. If a list is provided, the returned trajectory is an array of states corresponding to each element in the list.

PT2::ParameterizationType: The parameterization type for the τ2 parameter.

Keyword Arguments

manifold_length::AbstractFloat = 1.0: The length of the manifold, in CRTBP units. Sets the max length of the manifold, where τ2 = 1.0 corresponds to the full manifold_length.

solver = Vern9(): The DifferentialEquations.jl solver to use.

reltol = 1e-14: The relative tolerance for the solver.

abstol = 1e-14: The absolute tolerance for the solver.

CRTBPNaturalMotion.get_stable_manifold_trajectory — Method

get_stable_manifold_trajectory(
    man::InvariantManifold, left_pert::Bool,
    τ1::AbstractFloat, PT1::ParameterizationType,
    start_cond::Function, term_cond::Function, PT2::ParameterizationType;
    solver = Vern9(),
    reltol = 1e-14,
    abstol = 1e-14,
)

Returns a single stable manifold trajectory given scalar parameter τ1, where the starting state of the returned trajectory is determined through the satisfaction of start_cond and the final state is determined through the satisfaction of term_cond, where start_cond and term_cond are termination condition callback functions.

Arguments

man::InvariantManifold: The invariant manifold.

left_pert::Bool: If true, the perturbation is to the left of the periodic orbit.

τ1::AbstractFloat: The parameter for the initial condition on the periodic orbit.

PT1::ParameterizationType: The parameterization type for the τ1 parameter.

start_cond::Function: The starting condition callback function for the solver.

term_cond::Function: The termination condition callback function for the solver.

PT2::ParameterizationType: The parameterization type for the final condition on the manifold.

Keyword Arguments

solver = Vern9(): The DifferentialEquations.jl solver to use.

reltol = 1e-14: The relative tolerance for the solver.

abstol = 1e-14: The absolute tolerance for the solver.

CRTBPNaturalMotion.get_stable_manifold_trajectory — Method

get_stable_manifold_trajectory(
    man::InvariantManifold, left_pert::Bool,
    τ1::AbstractFloat,   PT1::ParameterizationType,
    term_cond::Function, PT2::ParameterizationType;
    solver = Vern9(),
    reltol = 1e-14,
    abstol = 1e-14,
)

Returns a single stable manifold trajectory given scalar parameter τ1, where the final state of the returned trajectory is determined through the satisfaction of term_cond, a termination condition callback function.

Arguments

man::InvariantManifold: The invariant manifold.

left_pert::Bool: If true, the perturbation is to the left of the periodic orbit.

τ1::AbstractFloat: The parameter for the initial condition on the periodic orbit.

PT1::ParameterizationType: The parameterization type for the τ1 parameter.

term_cond::Function: The termination condition callback function for the solver.

PT2::ParameterizationType: The parameterization type for the final condition on the manifold.

Keyword Arguments

solver = Vern9(): The DifferentialEquations.jl solver to use.

reltol = 1e-14: The relative tolerance for the solver.

abstol = 1e-14: The absolute tolerance for the solver.

CRTBPNaturalMotion.get_unstable_manifold_trajectory — Method

get_unstable_manifold_trajectory(
    man::InvariantManifold, left_pert::Bool,
    τ1::AbstractFloat, PT1::ParameterizationType,
    start_cond::Union{AbstractFloat,Function},
    τ2::Union{AbstractFloat,ParameterListType}, PT2::ParameterizationType;
    manifold_length::AbstractFloat = 1.0,
    solver = Vern9(),
    reltol = 1e-14,
    abstol = 1e-14,
)

Returns a single unstable manifold trajectory corresponding to the parameters τ1 and τ2. Here, the starting point of the returned trajectory is determined by the argument start_cond. If start_cond is a scalar, this specifies the distance from the initial condition on the periodic orbit in terms of PT2. If start_cond is a function, this specifies a termination condition callback function for the solver that will be used to propagate to the initial state of the returned trajectory from the initial condition on the periodic orbit.

Arguments

man::InvariantManifold: The invariant manifold.

left_pert::Bool: If true, the perturbation is to the left of the periodic orbit.

τ1::AbstractFloat: The parameter for the initial condition on the periodic orbit.

PT1::ParameterizationType: The parameterization type for the τ1 parameter.

start_cond::Union{AbstractFloat,Function}: The starting condition for the manifold trajectory. If a scalar, this specifies the distance from the initial condition on the periodic orbit in terms of PT2. If a function, this specifies a termination condition callback function for the solver that will be used to propagate to the initial state of the returned trajectory from the initial condition on the periodic orbit.

τ2::Union{AbstractFloat,ParameterListType}: The parameter for the final condition on the manifold. If a single scalar is provided, the returned trajectory is a DifferentialEquations.jl solution struct. If a list is provided, the returned trajectory is an array of states corresponding to each element in the list.

PT2::ParameterizationType: The parameterization type for the τ2 parameter.

Keyword Arguments

manifold_length::AbstractFloat = 1.0: The length of the manifold, in CRTBP units. Sets the max length of the manifold, where τ2 = 1.0 corresponds to the full manifold_length, starting from the state parameterized by start_cond.

solver = Vern9(): The DifferentialEquations.jl solver to use.

reltol = 1e-14: The relative tolerance for the solver.

abstol = 1e-14: The absolute tolerance for the solver.

CRTBPNaturalMotion.get_unstable_manifold_trajectory — Method

get_unstable_manifold_trajectory(
    man::InvariantManifold, left_pert::Bool,
    τ1::AbstractFloat, PT1::ParameterizationType,
    τ2::Union{AbstractFloat,ParameterListType}, PT2::ParameterizationType;
    manifold_length::AbstractFloat = 1.0,
    solver = Vern9(),
    reltol = 1e-14,
    abstol = 1e-14,
)

Returns a single unstable manifold trajectory corresponding to the parameters τ1 and τ2.

Arguments

man::InvariantManifold: The invariant manifold.

left_pert::Bool: If true, the perturbation is to the left of the periodic orbit.

τ1::AbstractFloat: The parameter for the initial condition on the periodic orbit.

PT1::ParameterizationType: The parameterization type for the τ1 parameter.

τ2::Union{AbstractFloat,ParameterListType}: The parameter for the final condition on the manifold. If a single scalar is provided, the returned trajectory is a DifferentialEquations.jl solution struct. If a list is provided, the returned trajectory is an array of states corresponding to each element in the list.

PT2::ParameterizationType: The parameterization type for the τ2 parameter.

Keyword Arguments

manifold_length::AbstractFloat = 1.0: The length of the manifold, in CRTBP units. Sets the max length of the manifold, where τ2 = 1.0 corresponds to the full manifold_length.

solver = Vern9(): The DifferentialEquations.jl solver to use.

reltol = 1e-14: The relative tolerance for the solver.

abstol = 1e-14: The absolute tolerance for the solver.

CRTBPNaturalMotion.get_unstable_manifold_trajectory — Method

get_unstable_manifold_trajectory(
    man::InvariantManifold, left_pert::Bool,
    τ1::AbstractFloat, PT1::ParameterizationType,
    start_cond::Function, term_cond::Function, PT2::ParameterizationType;
    solver = Vern9(),
    reltol = 1e-14,
    abstol = 1e-14,
)

Returns a single unstable manifold trajectory given scalar parameter τ1, where the starting state of the returned trajectory is determined through the satisfaction of start_cond and the final state is determined through the satisfaction of term_cond, where start_cond and term_cond are termination condition callback functions.

Arguments

man::InvariantManifold: The invariant manifold.

left_pert::Bool: If true, the perturbation is to the left of the periodic orbit.

τ1::AbstractFloat: The parameter for the initial condition on the periodic orbit.

PT1::ParameterizationType: The parameterization type for the τ1 parameter.

start_cond::Function: The starting condition callback function for the solver.

term_cond::Function: The termination condition callback function for the solver.

PT2::ParameterizationType: The parameterization type for the final condition on the manifold.

Keyword Arguments

solver = Vern9(): The DifferentialEquations.jl solver to use.

reltol = 1e-14: The relative tolerance for the solver.

abstol = 1e-14: The absolute tolerance for the solver.

CRTBPNaturalMotion.get_unstable_manifold_trajectory — Method

get_unstable_manifold_trajectory(
    man::InvariantManifold, left_pert::Bool,
    τ1::AbstractFloat,   PT1::ParameterizationType,
    term_cond::Function, PT2::ParameterizationType;
    solver = Vern9(),
    reltol = 1e-14,
    abstol = 1e-14,
)

Returns a single unstable manifold trajectory given scalar parameter τ1, where the final state of the returned trajectory is determined through the satisfaction of term_cond, a termination condition callback function.

Arguments

man::InvariantManifold: The invariant manifold.

left_pert::Bool: If true, the perturbation is to the left of the periodic orbit.

τ1::AbstractFloat: The parameter for the initial condition on the periodic orbit.

PT1::ParameterizationType: The parameterization type for the τ1 parameter.

term_cond::Function: The termination condition callback function for the solver.

PT2::ParameterizationType: The parameterization type for the final condition on the manifold.

Keyword Arguments

solver = Vern9(): The DifferentialEquations.jl solver to use.

reltol = 1e-14: The relative tolerance for the solver.

abstol = 1e-14: The absolute tolerance for the solver.

Type Flags

CRTBPNaturalMotion.ArcLength — Type

ArcLength

A concrete type to represent arc-length as the independant variable.

CRTBPNaturalMotion.Time — Type

Time

A concrete type to represent time as the independant variable.

Propagation Functions

CRTBPNaturalMotion.propagate_return_all_states — Method

propagate_return_all_states(
    x0::SVector{6,TX},
    tspan::Tuple{TT,TT},
    mu,
    cond::Function,
    affect::Function,
    iv::Type{AbstractIndependantVariable};
    solver = Vern9(),
    reltol = 1e-14,
    abstol = 1e-14,
) where {TX,TT}

Propagate the state x0 over the time span tspan and return the full ODE solution. This is done employing a DifferentialEquations.jl ContinuousCallback defined with cond and affect. The final argument can be specified to set the desired independant variable (i.e., Time or ArcLength).

Arguments

x0::SVector{6,TX}: The initial state with x0 = [r0; v0].

tspan::Tuple{TT,TT}: The span of the independant variable forwhich to solve the ODE.

mu::Real: The mass parameter of the three-body system (mu = m2 / (m1 + m2)).

cond::Function: The DifferentialEquations.jl continuous callback condition function that will apply affect to the integrator struct (see DifferentialEquations.jl callback documentation).

affect::Function: The DifferentialEquations.jl continuous callback affect function.

iv::Type{AbstractIndependantVariable}: The desired independant variable (Time if unspecified).

Keywords

solver::OrdinaryDiffEq.AbstractODESolver: The solver to use for the integration.

reltol: The relative tolerance for the solver.

abstol: The absolute tolerance for the solver.

Returns

ODESolution: The full solution to the ODE problem.

CRTBPNaturalMotion.propagate_return_all_states — Method

propagate_return_all_states(
    x0::SVector{6,TX},
    tspan::Tuple{TT,TT},
    mu,
    term_cond::Function[, iv::Type{AbstractIndependantVariable}];
    solver = Vern9(),
    reltol = 1e-14,
    abstol = 1e-14,
) where {TX,TT}

Propagate the state x0 over the time span tspan and return the full ODE solution. At the first point at which term_cond is zero (if any), the integration will terminate. The final argument can be specified to set the desired independant variable (i.e., Time or ArcLength).

Arguments

x0::SVector{6,TX}: The initial state with x0 = [r0; v0].

tspan::Tuple{TT,TT}: The span of the independant variable forwhich to solve the ODE.

mu::Real: The mass parameter of the three-body system (mu = m2 / (m1 + m2)).

term_cond::Function: The DifferentialEquations.jl continuous callback condition function that will terminate the integration when it is zero. Function should be of the form term_cond(x,t,integ), where x is the state, t is the independant variable, and integ is the integrator struct (see DifferentialEquations.jl callback documentation).

iv::Type{AbstractIndependantVariable}: The desired independant variable (Time if unspecified).

Keywords

solver::OrdinaryDiffEq.AbstractODESolver: The solver to use for the integration.

reltol: The relative tolerance for the solver.

abstol: The absolute tolerance for the solver.

Returns

ODESolution: The full solution to the ODE problem.

CRTBPNaturalMotion.propagate_return_all_states — Method

propagate_return_all_states(
    x0::SVector{6,TX}, tspan::Tuple{TT,TT}, mu [, iv::Type{AbstractIndependantVariable}];
    solver = Vern9(),
    reltol = 1e-14,
    abstol = 1e-14,
) where {TX,TT}

Propagate the state x0 over the time span tspan and return the full ODE solution. The final argument can be specified to set the desired independant variable (i.e., Time or ArcLength).

Arguments

x0::SVector{6,T}: The initial state with x0 = [r0; v0].

tspan::Tuple{T,T}: The span of the independant variable forwhich to solve the ODE.

mu::Real: The mass parameter of the three-body system (mu = m2 / (m1 + m2)).

iv::Type{AbstractIndependantVariable}: The desired independant variable (Time if unspecified).

Keywords

solver::OrdinaryDiffEq.AbstractODESolver: The solver to use for the integration.

reltol: The relative tolerance for the solver.

abstol: The absolute tolerance for the solver.

Returns

ODESolution: The full solution to the ODE problem.

CRTBPNaturalMotion.propagate_return_all_states — Method

propagate_return_all_states(
    x0::SVector{6,TX},
    tsteps::Union{StepRangeLen,AbstractVector},
    mu,
    iv::Type{AbstractIndependantVariable};
    solver = Vern9(),
    reltol = 1e-14,
    abstol = 1e-14,
) where TX

Propagate the state x0 from tsteps[1] to tsteps[end], returning the solution at each time in tsteps. The final argument can be specified to set the desired independant variable (i.e., Time or ArcLength).

Arguments

x0::SVector{6,TX}: The initial state with x0 = [r0; v0].

tsteps::Union{StepRangeLen,AbstractVector}: The time steps at which a state x(t) is desired.

mu::Real: The mass parameter of the three-body system (mu = m2 / (m1 + m2)).

iv::Type{AbstractIndependantVariable}: The desired independant variable.

Keywords

solver::OrdinaryDiffEq.AbstractODESolver: The solver to use for the integration.

reltol: The relative tolerance for the solver.

abstol: The absolute tolerance for the solver.

Returns

Vector{SVector{6,TX}}: A vector of states at each time in tsteps.

CRTBPNaturalMotion.propagate_return_arc_length — Method

propagate_return_arc_length(
    x0::SVector{6,TX}, tspan::Tuple{TT,TT}, mu;
    solver = Vern9(),
    reltol = 1e-14,
    abstol = 1e-14,
) where {TX,TT}

Propagate the state x0 over the time span tspan and return the arclength along the trajectory. The final argument can be specified to set the desired independant variable (i.e., Time or ArcLength).

arguments

x0::SVector{6,TX}: the initial state with x0 = [r0; v0].

tspan::Tuple{TT,TT}: the time span forwhich to solve the ode.

mu::Real: the mass parameter of the three-body system (mu = m2 / (m1 + m2)).

keywords

solver::OrdinaryDiffEq.AbstractODESolver: The solver to use for the integration.

reltol: The relative tolerance for the solver.

abstol: The absolute tolerance for the solver.

returns

TX: The arc-length along the curve

CRTBPNaturalMotion.propagate_return_final_state — Method

propagate_return_final_state(
    x0::SVector{6,TX},
    tspan::Tuple{TT,TT},
    mu,
    term_cond::Function [, iv::Type{AbstractIndependantVariable}];
    solver = Vern9(),
    reltol = 1e-14,
    abstol = 1e-14,
) where {TX,TT}

Propagate the state x0 over the time span tspan and return the final state. At the first point at which term_cond is zero (if any), the integration will terminate. The final argument can be specified to set the desired independant variable (i.e., Time or ArcLength).

Arguments

x0::SVector{6,TX}: The initial state with x0 = [r0; v0].

tspan::Tuple{TT,TT}: The span of the independant variable forwhich to solve the ODE.

mu::Real: The mass parameter of the three-body system (mu = m2 / (m1 + m2)).

term_cond::Function: The DifferentialEquations.jl continuous callback condition function that will terminate the integration when it is zero. Function should be of the form term_cond(x,t,integ), where x is the state, t is the independant variable, and integ is the integrator struct (see DifferentialEquations.jl callback documentation).

iv::Type{AbstractIndependantVariable}: The desired independant variable (Time if unspecified).

Keywords

solver::OrdinaryDiffEq.AbstractODESolver: The solver to use for the integration.

reltol: The relative tolerance for the solver.

abstol: The absolute tolerance for the solver.

Returns

SVector{6,TX}: The final state of the ODE solution.

CRTBPNaturalMotion.propagate_return_final_state — Method

propagate_return_final_state(
    x0::SVector{6,TX}, tspan::Tuple{TT,TT}, mu [, iv::Type{AbstractIndependantVariable}];
    solver = Vern9(),
    reltol = 1e-14,
    abstol = 1e-14,
) where {TX,TT}

Propagate the state x0 over the time span tspan and return the final state. The final argument can be specified to set the desired independant variable (i.e., Time or ArcLength).

arguments

x0::SVector{6,TX}: The initial state with x0 = [r0; v0].

tspan::Tuple{TT,TT}: The span of the independant variable forwhich to solve the ode.

mu::Real: The mass parameter of the three-body system (mu = m2 / (m1 + m2)).

iv::Type{AbstractIndependantVariable}: The desired independant variable (Time if unspecified).

keywords

solver::OrdinaryDiffEq.AbstractODESolver: The solver to use for the integration.

reltol: The relative tolerance for the solver.

abstol: The absolute tolerance for the solver.

returns

SVector{6,TX}: The final state of the ode solution.

CRTBPNaturalMotion.propagate_return_final_stm — Method

propagate_return_final_stm(
    x0::SVector{6,TX}, tspan::Tuple{TT,TT}, mu;
    solver = Vern9(),
    reltol = 1e-14,
    abstol = 1e-14,
) where {TX,TT}

Propagate the state x0 over the time span tspan and return the final state transition matrix (STM).

arguments

x0::SVector{6,TX}: The initial state with x0 = [r0; v0].

tspan::Tuple{TT,TT}: The span of the independant variable forwhich to solve the ode.

mu::Real: The mass parameter of the three-body system (mu = m2 / (m1 + m2)).

keywords

solver::OrdinaryDiffEq.AbstractODESolver: The solver to use for the integration.

reltol: The relative tolerance for the solver.

abstol: The absolute tolerance for the solver.

returns

SMatrix{6,6,TX,36}: The stm for the full trajectory

CRTBPNaturalMotion.propagate_return_time — Method

propagate_return_time(
    x0::SVector{6,TX}, sspan::Tuple{TT,TT}, mu;
    solver = Vern9(),
    reltol = 1e-14,
    abstol = 1e-14,
) where {TX,TT}

Propagate the state x0 over the arc-length span sspan and return the arclength along the trajectory. The final argument can be specified to set the desired independant variable (i.e., Time or ArcLength).

arguments

x0::SVector{6,TX}: the initial state with x0 = [r0; v0].

sspan::Tuple{TT,TT}: the arc-length span forwhich to solve the ode.

mu::Real: the mass parameter of the three-body system (mu = m2 / (m1 + m2)).

keywords

solver::OrdinaryDiffEq.AbstractODESolver: The solver to use for the integration.

reltol: The relative tolerance for the solver.

abstol: The absolute tolerance for the solver.

returns

TX: The time along the curve

Utility Functions

CRTBPNaturalMotion.jacobi_integral — Method

jacobi_integral(x, μ)

Computes the Jacobi integral for the CRTBP given the state x and mass parameter μ.

Arguments

x::AbstractVector: State vector.

μ::Real: Mass parameter.

CRTBPNaturalMotion.jacobian — Method

jacobian(m::FastChebInterpolation{IT}, τ1, τ2) where IT -> SMatrix{6, 2, Float64, 12}

Evaluate the Jacobian of the interpolation at the given τ1 and τ2 for a bi-variate interpolation.

Arguments

m::FastChebInterpolation{IT}: The interpolation object.

τ1: The first variable.

τ2: The second variable.

Returns

SMatrix{6, 2, Float64, 12}: The Jacobian of the interpolation.

CRTBPNaturalMotion.jacobian — Method

jacobian(m::FastChebInterpolation{IT}, τ1) where IT -> SVector{6, Float64}

Evaluate the Jacobian of the interpolation at the given τ1 for a uni-variate interpolation.

Arguments

m::FastChebInterpolation{IT}: The interpolation object.

τ1: The variable.

Returns

SVector{6, Float64}: The Jacobian of the interpolation.

CRTBPNaturalMotion.value — Method

value(m::FastChebInterpolation{IT}, τ1[, τ2]) where IT -> SVector{6, Float64}

Evaluate the interpolation at the given τ1 and τ2 (if employing a bi-variate interpolation).

Arguments

m::FastChebInterpolation{IT}: The interpolation object.

τ1: The variable.

τ2: The second variable (if employing a bi-variate interpolation).

Returns

SVector{6, Float64}: The interpolated state.